+++ src/main/java/net/soulsweaponry/util/ModTags.java
@@ -32,6 +32,7 @@ public class ModTags {
         public static final TagKey<Item> TRICK_WEAPONS = createTag("trick_weapons");
         public static final TagKey<Item> GUNS = createTag("guns");
         public static final TagKey<Item> LOST_SOUL_REPAIR = createTag("lost_soul_items_repair_ingredients");
+        public static final TagKey<Item> DRAGONBANE_MATERIAL = createTag("dragonbane_material");
 
         public static final TagKey<Item> STICKS = createCommonTag("wood_sticks");
         public static final TagKey<Item> SILVER_INGOTS = createCommonTag("silver_ingots");